refactor(domain): migrate credential storage to domain layer
