+++ cloud-wallet-openid4vc/Cargo.toml
@@ -6,13 +6,10 @@ license.workspace = true
 
 [dependencies]
 cloud-wallet-crypto = { path = "../cloud-wallet-crypto", features = ["jwk"] }
-cloud-wallet-kms = { path = "../cloud-wallet-kms", optional = true }
 
-async-trait.workspace = true
 base64.workspace = true
 color-eyre.workspace = true
 csscolorparser = { version = "0.8", features = ["serde"] }
-dashmap = { workspace = true, optional = true }
 futures = "0.3"
 jsonwebtoken = { workspace = true, features = ["aws_lc_rs"] }
 percent-encoding = "2"
@@ -26,21 +23,8 @@ serde_with.workspace = true
 thiserror.workspace = true
 time.workspace = true
 url = { workspace = true, features = ["serde"] }
-uuid = { workspace = true, features = ["serde", "v4"] }
 validator = { version = "0.20", features = ["derive"] }
-
-[dependencies.sqlx]
-workspace = true
-features = ["runtime-tokio", "tls-rustls-aws-lc-rs"]
-optional = true
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 wiremock = "0.6"
-
-[features]
-default = ["memory-repo"]
-memory-repo = ["dep:cloud-wallet-kms", "dep:dashmap"]
-mysql = ["dep:cloud-wallet-kms", "dep:sqlx", "sqlx/mysql"]
-postgres = ["dep:cloud-wallet-kms", "dep:sqlx", "sqlx/postgres"]
-sqlite = ["dep:cloud-wallet-kms", "dep:sqlx", "sqlx/sqlite"]
