+++ cloud-wallet-openid4vc/src/lib.rs
@@ -1,5 +1,3 @@
-pub mod credential;
 pub mod errors;
 pub mod http;
 pub mod issuance;
-pub mod storage;
